Renamed package 'resource' to 'resources' as per DropWizard convention; made package 'auth' a subpackage of 'core' as per same
