+++ src/main/java/nl/knaw/dans/sword2/DdSword2Application.java
@@ -24,8 +24,8 @@ import io.dropwizard.client.HttpClientBuilder;
 import io.dropwizard.forms.MultiPartBundle;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
-import nl.knaw.dans.sword2.auth.Depositor;
-import nl.knaw.dans.sword2.auth.SwordAuthenticator;
+import nl.knaw.dans.sword2.core.auth.Depositor;
+import nl.knaw.dans.sword2.core.auth.SwordAuthenticator;
 import nl.knaw.dans.sword2.core.finalizer.DepositFinalizerEvent;
 import nl.knaw.dans.sword2.core.finalizer.DepositFinalizerManager;
 import nl.knaw.dans.sword2.core.service.BagExtractorImpl;
@@ -45,11 +45,11 @@ import nl.knaw.dans.sword2.health.ExecutorQueueHealthCheck;
 import nl.knaw.dans.sword2.health.FileSystemPermissionHealthCheck;
 import nl.knaw.dans.sword2.health.QueueHealthCheck;
 import nl.knaw.dans.sword2.health.UploadDepositOnSameFileSystemHealthCheck;
-import nl.knaw.dans.sword2.resource.CollectionResourceImpl;
-import nl.knaw.dans.sword2.resource.ContainerResourceImpl;
-import nl.knaw.dans.sword2.resource.HashHeaderInterceptor;
-import nl.knaw.dans.sword2.resource.ServiceDocumentResourceImpl;
-import nl.knaw.dans.sword2.resource.StatementResourceImpl;
+import nl.knaw.dans.sword2.resources.CollectionResourceImpl;
+import nl.knaw.dans.sword2.resources.ContainerResourceImpl;
+import nl.knaw.dans.sword2.resources.HashHeaderInterceptor;
+import nl.knaw.dans.sword2.resources.ServiceDocumentResourceImpl;
+import nl.knaw.dans.sword2.resources.StatementResourceImpl;
 import org.glassfish.jersey.media.multipart.MultiPartFeature;
 
 import java.util.concurrent.ArrayBlockingQueue;
