+++ src/main/java/nl/knaw/dans/sword2/core/auth/Depositor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.auth;
+package nl.knaw.dans.sword2.core.auth;
 
 import java.security.Principal;
 import java.util.Set;
