+++ src/main/java/nl/knaw/dans/sword2/resources/ContainerResource.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.resource;
+package nl.knaw.dans.sword2.resources;
 
 import io.dropwizard.auth.Auth;
-import nl.knaw.dans.sword2.auth.Depositor;
+import nl.knaw.dans.sword2.core.auth.Depositor;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.HEAD;
