+++ src/main/java/nl/knaw/dans/sword2/resources/HashHeaderInterceptor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.resource;
+package nl.knaw.dans.sword2.resources;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
