+++ src/main/java/nl/knaw/dans/sword2/resources/ServiceDocumentResource.java
@@ -13,25 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.resource;
+package nl.knaw.dans.sword2.resources;
 
 import io.dropwizard.auth.Auth;
-import nl.knaw.dans.sword2.auth.Depositor;
+import nl.knaw.dans.sword2.core.auth.Depositor;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-@Path("/statement/{id}")
-public interface StatementResource {
+@Path("/servicedocument")
+public interface ServiceDocumentResource {
+
     @GET
-    @Produces(MediaType.APPLICATION_XML)
-    Response getStatement(@PathParam("id") String depositId, @Context HttpHeaders headers, @Auth Depositor depositor);
-
-
+    @Produces(MediaType.APPLICATION_ATOM_XML)
+    Response getServiceDocument(@Context HttpHeaders httpHeaders, @Auth Depositor depositor);
 }
