+++ src/main/java/nl/knaw/dans/sword2/resources/ServiceDocumentResourceImpl.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.resource;
+package nl.knaw.dans.sword2.resources;
 
 import nl.knaw.dans.sword2.core.config.UriRegistry;
-import nl.knaw.dans.sword2.auth.Depositor;
+import nl.knaw.dans.sword2.core.auth.Depositor;
 import nl.knaw.dans.sword2.core.config.CollectionConfig;
 import nl.knaw.dans.sword2.api.service.ServiceCollection;
 import nl.knaw.dans.sword2.api.service.ServiceDocument;
