+++ src/test/java/nl/knaw/dans/sword2/auth/SwordAuthenticatorTest.java
@@ -17,6 +17,7 @@ package nl.knaw.dans.sword2.auth;
 
 import io.dropwizard.auth.AuthenticationException;
 import io.dropwizard.auth.basic.BasicCredentials;
+import nl.knaw.dans.sword2.core.auth.SwordAuthenticator;
 import nl.knaw.dans.sword2.core.config.UserConfig;
 import org.apache.http.ProtocolVersion;
 import org.apache.http.client.HttpClient;
